requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pymongo>=4.6.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pydantic>=2.8.0
boto3>=1.34.0
neo4j>=5.11.0
requests>=2.31.0
# pip install -r requirements.txt